Fixed autoinstall cleanup on machines that globally disable `package-lock.json`

+++ src/_cleanup.js
@@ -1,23 +1,23 @@
-let { mkdirSync, renameSync } = require('fs')
+let { existsSync, mkdirSync, renameSync } = require('fs')
 let { join } = require('path')
 
 // Best effort local artifact cleanup
 module.exports = function cleanup (installed) {
   if (installed) {
-    try {
-      installed.forEach(i => {
-        let { dir, file, remove } = i
+    installed.forEach(i => {
+      let { dir, file, remove } = i
+      try {
         if (file === 'package.json') {
           let dest = join(dir, 'node_modules', '_arc-autoinstall')
           mkdirSync(dest, { recursive: true })
           remove.forEach(f => {
             let before = join(dir, f)
             let after = join(dest, f)
-            renameSync(before, after)
+            if (existsSync(before)) renameSync(before, after)
           })
         }
-      })
-    }
-    catch (err) { null } // Swallow errors, we may have to bubble something else
+      }
+      catch (err) { null } // Swallow errors, we may have to bubble something else
+    })
   }
 }
